Add files via upload
This project focuses on the application of Convolutional Neural Networks (CNN) and Transfer Learning techniques to classify malaria from cell images. In this project, we develop a CNN model by using Transfer Learning to leverage knowledge from pre-trained networks with a dataset specifically compiled for malaria-infected and uninfected blood smear images. The objective is to accurately distinguish between infected and uninfected cells, thereby assisting in rapid and efficient malaria diagnosis.